修改接口逻辑

+++ Locker_Services/src/main/java/cn/lingnan/dto/LockerBasicInformation.java
@@ -1,11 +1,12 @@
 package cn.lingnan.dto;
 
-import java.util.Date;
 import lombok.Data;
+
+import java.util.Date;
 
 @Data
 public class LockerBasicInformation {
-    private Integer lockerId;
+    private Long lockerId;
 
     private String province;
